requirements/prod.txt
# Everything needed in production

# Flask
click>=7.0
Flask==3.1.1
Werkzeug==3.1.3

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41

# Migrations
Flask-Migrate==4.1.0

# Forms
email-validator==2.2.0
Flask-WTF==1.2.2
WTForms==3.2.1

# Deployment
gevent==25.5.1
gunicorn>=19.9.0
supervisor==4.2.5

# Flask Static Digest
Flask-Static-Digest==0.4.1

# Auth
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3

# Caching
Flask-Caching>=2.0.2

# Debug toolbar
Flask-DebugToolbar==0.16.0

# Environment variable parsing
environs==14.2.0
Flask-Mail>=0.9.1

Flask-CORS==4.0.0

Flask-JWT-Extended==4.6.0


boto3==1.34.122

openai
google-generativeai

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
webtest>=3.0.0
factory_boy>=3.3.0

python-docx==1.1.2
bibtexparser>=1.4

